+++ src/SearchSystem.UserInteraction/SearchSystem.UserInteraction.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="$(MsBuildThisFileDirectory)\..\common.props" />
+
+	<ItemGroup>
+		<ProjectReference Include="..\SearchSystem.Infrastructure\SearchSystem.Infrastructure.csproj" />
+	</ItemGroup>
+
+</Project>
