+++ src/SearchSystem.sln
@@ -21,6 +21,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SearchSystem.BooleanSearch"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SearchSystem.VectorSearch", "SearchSystem.VectorSearch\SearchSystem.VectorSearch.csproj", "{758D5E1A-58B6-4C91-979B-AE92D2AF14DE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SearchSystem.UserInteraction", "SearchSystem.UserInteraction\SearchSystem.UserInteraction.csproj", "{CCB9EB7E-EA0E-4636-9FBA-7D3FC00B1F7B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -59,5 +61,9 @@ Global
 		{758D5E1A-58B6-4C91-979B-AE92D2AF14DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{758D5E1A-58B6-4C91-979B-AE92D2AF14DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{758D5E1A-58B6-4C91-979B-AE92D2AF14DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CCB9EB7E-EA0E-4636-9FBA-7D3FC00B1F7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CCB9EB7E-EA0E-4636-9FBA-7D3FC00B1F7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CCB9EB7E-EA0E-4636-9FBA-7D3FC00B1F7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CCB9EB7E-EA0E-4636-9FBA-7D3FC00B1F7B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
